fix totalsegmentator v1
When installing totalsegmentator pinned to v1.5.6 we get the following dependency resolution: totalsegmentator v1.5.6 > nnunet-customized v1.2 > torch v2.6.0.

There seems to be a breaking change introduced in torch v2.6.0: In PyTorch 2.6, we changed the default value of the weights_only argument in torch.load from False to True.

This causes the weights to not load properly.

We therefore manually pin torch to 2.2.2 (which is the latest version that was tested by the MHub team).